requirements/dev.txt
anyascii==0.3.2
asgiref==3.6.0
beautifulsoup4==4.11.2
black==23.3.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==41.0.1
defusedxml==0.7.1
dj-rest-auth==4.0.1
Django==4.2.1
django-allauth==0.54.0
django-cors-headers==4.0.0
django-debug-toolbar==4.1.0
django-filter<24,>=23.3
django-modelcluster>=6.1,<7.0
django-permissionedforms==0.1
django-taggit<5.1,>=4.0
django-treebeard==4.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
filetype==1.2.0
gunicorn==20.1.0
html5lib==1.1
idna==3.4
l18n==2021.3
mypy-extensions==1.0.0
oauthlib==3.2.2
openpyxl==3.1.2
packaging==23.1
pathspec==0.11.1
Pillow==9.5.0
pillow-heif==0.13.0
platformdirs==3.5.1
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.7.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
requests==2.30.0
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
telepath>=0.3.1,<1
tomli==2.0.1
urllib3==2.0.2
wagtail>=6.0.2,<6.0.3
wagtail-headless-preview==0.5.0
webencodings==0.5.1
Willow==1.6.2